// candidatos-ui/src/app/core/models/report-params..model.ts
// Interface para los parámetros del reporte Certificado de otacion
export interface ReportParams {
  votoLiteral: string;
  nombrecompleto: string;
  documento: string;
  mesa: string;
  recinto: string;
  pais: string;
  departamento: string;
  localidad: string;
  fecha: string;
}